requirements.txt
playwright
requests
python-dotenv
pytest
openai